app/requirements.txt
asgiref==3.3.4
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
decorator==5.0.9
fastapi==0.65.1
google-api-core==1.29.0
google-api-python-client==2.7.0
google-auth==1.30.1
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
h11==0.12.0
httplib2==0.19.1
httptools==0.2.0
idna==2.10
packaging==20.9
protobuf==3.17.2
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.8.2
pymongo==3.11.4
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
retry==0.9.2
retrying==1.3.3
rsa==4.7.2
six==1.16.0
starlette==0.14.2
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.5
uvicorn==0.14.0
uvloop==0.15.2
watchgod==0.7
websockets==9.1
